# Fix: visualizer performance