--- a/src/modules/simulator/index.ts
+++ b/src/modules/simulator/index.ts
@@ -9,6 +9,11 @@ class Simulator {
   private scrollEndCount: number;
   private touchScrollPrevY?: number;
   private scrollY: number;
+  private animationId?: number;
+  private lastFrameTime = 0;
+  private targetFrameRate = 60;
+  private isVisible = true;
+  private isPaused = false;
 
   public constructor(visualizer: Visualizer) {
     this.visualizer = visualizer;
@@ -18,14 +23,56 @@ class Simulator {
       const query = new URL(document.location.href).searchParams;
       return Number.parseInt(query.get("depth") || "0", 10);
     })();
+
+    document.addEventListener("visibilitychange", this.handleVisibilityChange);
   }
 
-  public start = (framerate: number): void => {
-    const refreshMs = 1000 / framerate;
-    setInterval(() => {
+  public start = (framerate = 60): void => {
+    this.targetFrameRate = framerate;
+    this.animate();
+  };
+
+  private animate = (currentTime: number = performance.now()): void => {
+    if (!this.isVisible || this.isPaused) {
+      this.animationId = requestAnimationFrame(this.animate);
+      return;
+    }
+
+    const deltaTime = currentTime - this.lastFrameTime;
+    const targetInterval = 1000 / this.targetFrameRate;
+
+    if (deltaTime >= targetInterval) {
       const calcResult = this.simulate();
       this.visualizer.draw(calcResult);
-    }, refreshMs);
+      this.lastFrameTime = currentTime;
+    }
+
+    this.animationId = requestAnimationFrame(this.animate);
+  };
+
+  public stop = (): void => {
+    if (this.animationId) {
+      cancelAnimationFrame(this.animationId);
+      this.animationId = undefined;
+    }
+  };
+
+  public pause = (): void => {
+    this.isPaused = true;
+  };
+
+  public resume = (): void => {
+    this.isPaused = false;
+    if (!this.animationId) {
+      this.animate();
+    }
+  };
+
+  private handleVisibilityChange = (): void => {
+    this.isVisible = !document.hidden;
+    if (this.isVisible && !this.animationId) {
+      this.animate();
+    }
   };
 
   private restrictScroll = () => {